renamed mmr.rs to history_tree.rs
